Added .agenda (also to .ignore), fixed oversight with appending to self.boxes in RoomMapper.
At the end of select, self.boxes.append(roombox) was being called regardless of whether or not a saved room was selected. This would cause self.boxes to be appended to every time a saved room was selected.

Also added Project.agenda for keeping track of what needs to be done and what bugs need to be fixed. Possibly will be moved to the GitHub page?